Implement basic grid layout
I figured getting the layout right would be the biggest rendering
task involved in the prototype, so I decided to tackle that first.

The requirements state:
> "The gameplay view consists of square cards laid out in a grid"

However the assets are rectangular. For the sake of simplicity I
decided to use the aspect ratio of the provided assets. This would
need discussed with stakeholders on a real project in case "square"
was more than a typo.

I decided to always show the back button at a constant size and
location (top left), since navigation is so important. For the
prototype I'm ignoring any localization concerns, such as
moving and flipping the back button for localizations that prefer
right-to-left languages. I suppose choice of asset for cards may
be culturally sensitive as well (e.g. perhaps some animals aren't
well known to children everywhere in the world), although for the
prototype I'll also ignore that once I get to gameplay.

With this prototype I notice that some grids lay out nicely in
some aspect ratios, but that my choice to give the button its own
top padding with an equal amount on the bottom leads to some grids
looking like they have a lot of unused space.